etl_web_portal/requirements.txt
blinker==1.9.0
click==8.3.0
colorama==0.4.6
et_xmlfile==2.0.0
Flask==3.1.2
greenlet==3.2.4
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mysql-connector-python==9.4.0
numpy==2.3.3
openpyxl==3.1.5
pandas==2.3.2
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.43
typing_extensions==4.15.0
tzdata==2025.2
Werkzeug==3.1.3
